Split unused file, tag and cluster entity deletion into a SELECT ... FOR UPDATE and a DELETE to prevent database locking

+++ genie-web/src/main/java/com/netflix/genie/web/services/ClusterService.java
@@ -393,5 +393,5 @@ public interface ClusterService {
      *
      * @return The number of clusters deleted
      */
-    int deleteTerminatedClusters();
+    long deleteTerminatedClusters();
 }
